Move all scope parsing to 'experimental'
Scope parsing is now exposed via `Scope.parse`, which is only
available in `globus_sdk.experimental.scope_parser` as a method of
parsing scopes.

The scope parsing passes through two IRs -- first a tree, then a graph
-- which reject cycles, normalize any optional/non-optional
redundancies, and ensure that the tree is built in the same way that
it will be during authorization flows in Globus Auth.

This replaces our "shadow" implementation of MutableScope, which had
been left in `scopes.scope_definition` under the name `Scope` but was
never exported or documented. All of the features of `Scope` under the
old location are now integrated into
`globus_sdk.experimental.scope_parser.Scope`.

Scope parsing is tested independently of other scope tests, moving and
expanding these tests into `tests/unit/experimental/`.
