+++ src/globus_sdk/experimental/scope_parser/__init__.py
@@ -1,0 +1,8 @@
+from .parser import ScopeCycleError, ScopeParseError
+from .scope_definition import Scope
+
+__all__ = (
+    "Scope",
+    "ScopeParseError",
+    "ScopeCycleError",
+)
